requirements.txt
<<<<<<< HEAD
autopep8==1.5.2
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
future==0.18.2
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
=======
certifi==2019.3.9
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
requests==2.21.0
urllib3==1.24.2
Werkzeug==0.15.2
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
requests==2.21.0
urllib3==1.24.2
Werkzeug==0.15.2
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
requests==2.21.0
urllib3==1.24.2
Werkzeug==0.15.2
>>>>>>> 895a1b0751182ddfeba6c44af5e242901bd287d3
